requirements.txt
pyaudio
google-cloud-speech
google-cloud-translate==2.0.1
tkinter